[ningpop] : edit page permission, many to many function

+++ funding/views.py
@@ -3,6 +3,7 @@ from django.views.decorators.http import require_POST
 
 from .models import Funding
 from write.models import Post
+from user.models import User
 from user.remind_module import sms_reminder
 
 import json
@@ -85,6 +86,9 @@ def funding_detail(request, funding_id):
 def funding_update(request, funding_id):
     funding = get_object_or_404(Funding, pk=funding_id)
 
+    if request.user != funding.name:
+        return redirect('funding_detail', funding.id)
+
     if request.method == 'POST':
         funding.community = request.POST['community']
         funding.community_address = request.POST['communityAddress']
@@ -102,6 +106,10 @@ def funding_update(request, funding_id):
 '''
 def funding_delete(request, funding_id):
     funding = get_object_or_404(Funding, pk=funding_id)
+    
+    if request.user != funding.name:
+        return redirect('funding_detail', funding.id)
+
     funding.delete()
     return redirect('funding_list')
 
@@ -117,13 +125,19 @@ def funding_counter(request):
     user_id = request.POST['user_id']
 
     funding = get_object_or_404(Funding, pk=funding_id)
+    user = get_object_or_404(User, pk=user_id)
 
-    result = funding.funding_counter() # 해당 펀딩 글의 펀딩 카운트 1 증가
-    if result != -1:
-        message = 'Success'
-        sms_reminder('funding', user_id, funding.title)
+    if funding.user.filter(id=user.id).exists():
+        message = 'exists'
     else:
-        message = 'Fail'
+        result = funding.funding_counter() # 해당 펀딩 글의 펀딩 카운트 1 증가
+        funding.user.add(user)
+        if result != -1:
+            message = 'Success'
+            sms_reminder('funding', user_id, funding.title)
+        else:
+            message = 'Fail'
+            sms_reminder('closed', user_id, funding.title)
 
     context = {
         "message": message,
